fix: complete provider-neutral check ordering remediation
